+++ gudp/src/service/conf.rs
@@ -5,11 +5,11 @@ pub struct Conf {
   pub example: usize,
 
   // Called when the packet is sent over the wire, with its sequence number
-  pub on_packet_sent: Option<Box<dyn FnMut((SocketAddr, SocketAddr), &[u8], usize) + Send>>,
+  pub on_packet_sent: Option<Box<dyn FnMut((SocketAddr, SocketAddr), &[u8], u32) + Send>>,
 
   // Called when the given sequence number is acked
-  pub on_packet_acked: Option<Box<dyn FnMut((SocketAddr, SocketAddr), usize) + Send>>,
+  pub on_packet_acked: Option<Box<dyn FnMut((SocketAddr, SocketAddr), u32) + Send>>,
 
   // Called when the given sequence number is lost (never acked and too old)
-  pub on_packet_lost: Option<Box<dyn FnMut((SocketAddr, SocketAddr), usize) + Send>>
+  pub on_packet_lost: Option<Box<dyn FnMut((SocketAddr, SocketAddr), u32) + Send>>
 }
